requirements.txt
fastapi 
uvicorn 
groq 
python-dotenv
chromadb
sentence-transformers
streamlit
